circular dependency fix by moving IMainThreadRunner to ChainSafe.Gaming
