+++ src/ChainSafe.Gaming.InProcessTransactionExecutor.Unity/RpcClientWrapper.cs
@@ -1,6 +1,6 @@
 using System;
-using ChainSafe.Gaming.Unity;
 using ChainSafe.Gaming.Web3;
+using ChainSafe.Gaming.Web3.Environment;
 using Nethereum.JsonRpc.Client;
 using Nethereum.Unity.Rpc;
 
@@ -10,7 +10,7 @@ namespace ChainSafe.Gaming.InProcessTransactionExecutor.Unity
     {
         public RpcClientWrapper(IChainConfig chainConfig, IMainThreadRunner mainThreadRunner)
         {
-            mainThreadRunner.Enqueue(delegate
+            mainThreadRunner.Enqueue(() =>
             {
                 Client = new UnityWebRequestRpcTaskClient(new Uri(chainConfig.Rpc));
             });
